# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Snow School Scheduler API
  description: スキー・スノーボードスクールのシフト管理システム API
  version: 1.0.0
  contact:
    name: Snow School Scheduler
servers:
  - url: http://localhost:3000
    description: 開発環境
  - url: https://snow-school-scheduler-api.example.com
    description: 本番環境
paths:
...
  /api/shifts:
    get:
      summary: シフト一覧取得
      description: シフトの一覧を取得（部門・種類・期間絞り込み可能）
      tags:
        - シフト管理
      parameters:
        - name: departmentId
          in: query
          description: 部門IDでフィルタリング
          schema:
            type: integer
        - name: shiftTypeId
          in: query
          description: シフト種類IDでフィルタリング
          schema:
            type: integer
        - name: dateFrom
          in: query
          description: 開始日でフィルタリング
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: 終了日でフィルタリング
          schema:
            type: string
            format: date
      responses:
        '200':
          description: シフト一覧
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ShiftWithStats'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: シフト作成
      description: 新しいシフトを作成
      tags:
        - シフト管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftInput'
      responses:
        '201':
          description: シフト作成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShiftWithStats'
                      message:
                        type: string
                        example: 'Shift operation completed successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
...
  responses:
...
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Validation failed'
    ServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Internal server error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InstructorSummary:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
          example: '山田'
        firstName:
          type: string
          example: '太郎'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, RETIRED]
          default: ACTIVE
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
          example: null
    ApiListResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        count:
          type: integer
          example: 10
        error:
          type: string
          nullable: true
          example: null
    ApiErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid invitation token'

    # 部門関連
    Department:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: 'ski'
        name:
          type: string
          example: 'スキー'
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ShiftType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'レッスン'
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        departmentId:
          type: integer
        shiftTypeId:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        department:
          $ref: '#/components/schemas/Department'
        shiftType:
          $ref: '#/components/schemas/ShiftType'
    ShiftInput:
      type: object
      required:
        - date
        - departmentId
        - shiftTypeId
      properties:
        date:
          type: string
          format: date
        departmentId:
          type: integer
        shiftTypeId:
          type: integer
        description:
          type: string
          nullable: true
        assignedInstructorIds:
          type: array
          items:
            type: integer
          description: '割り当てるインストラクターIDの配列'
          example: [1, 2, 3]
    ShiftWithStats:
      allOf:
        - $ref: '#/components/schemas/Shift'
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: '#/components/schemas/ShiftAssignment'
            assignedCount:
              type: integer
              description: 現在割り当てられているインストラクター数

    # シフト割り当て関連
    ShiftAssignment:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        instructorId:
          type: integer
        assignedAt:
          type: string
          format: date-time
        instructor:
          $ref: '#/components/schemas/InstructorSummary'
...